Refactor board and spec
Use `private` keyword to indicate which methods are truly private.

Move initialization code into `initialize`. It shouldn't be in a
separate method. If an object requires anything more than trivial
initialization, the object is doing way to much and needs to be
simplified.

Use standard code formatting.

Remove comments from spec. They don't add anything to the code and are a
distraction. By using git, whatever changes you have are always in the
history. If you want to return to some older version of the code, the
history is always there to help you get back there.

Also, define `winners` as a constant rather than a method. Since the
`winners` array doesn't ever change, it's a perfect use for a constant.
By using a method that returns an array, you have a situation where Ruby
will make a new array every time you refer to `winners`. Granted, this
isn't a huge problem, but it's still better as a constant.