Simplify board and spec
First, use instance doubles for players. The board object only depends
on the players' "mark" and so an instance double is a perfect
substitution.

Also, remove reference to `Game`. If the board object has no need to
refer to game, there is no need to have it in the spec, or in the
initializer.